refactor!: made BitAnd/Or/Not traits visible everywhere, not only in tests
